refactor(all): capitalize folder names, changes in all files checked.
